Run environment only once for all integration tests
